requirements.txt
Flask==2.3.2
python-telegram-bot==20.3
pandas
numpy
xgboost
plotly
requests
python-dotenv
